seperation into mulitple files: geometric, connectivty and mesh creation
