+++ CMakePresets.json
@@ -47,6 +47,7 @@
         "NeoN_BUILD_TESTS": "ON",
         "NeoN_BUILD_BENCHMARKS": "OFF",
         "NeoN_ENABLE_WARNINGS": "ON",
+        "NeoN_BUILD_PYTHON_BINDINGS": "ON",
         "Kokkos_ENABLE_DEBUG": "ON",
         "Kokkos_ENABLE_DEBUG_BOUNDS_CHECK": "ON",
         "Kokkos_ENABLE_COMPILE_AS_CMAKE_LANGUAGE": "OFF"
