+++ src/mesh/unstructured/createSingleCellMesh.cpp
@@ -1,0 +1,54 @@
+// SPDX-FileCopyrightText: 2023 - 2026 NeoN authors
+//
+// SPDX-License-Identifier: MIT
+
+#include "NeoN/mesh/unstructured/unstructuredMesh.hpp"
+
+namespace NeoN
+{
+
+UnstructuredMesh createSingleCellMesh(const Executor exec)
+{
+    // a 2D mesh in 3D space with left, right, top, bottom boundary faces
+    // with the centre at (0.5, 0.5, 0.0)
+    // left, top, right, bottom faces
+    // and four boundaries one left, right, top, bottom
+
+    vectorVector faceAreasVec3s(exec, {{-1, 0, 0}, {0, 1, 0}, {1, 0, 0}, {0, -1, 0}});
+    vectorVector faceCentresVec3s(
+        exec, {{0.0, 0.5, 0.0}, {0.5, 1.0, 0.0}, {1.0, 0.5, 0.0}, {0.5, 0.0, 0.0}}
+    );
+    scalarVector magFaceAreas(exec, {1, 1, 1, 1});
+
+    BoundaryMesh boundaryMesh(
+        exec,
+        {exec, {0, 0, 0, 0}},                                                           // faceCells
+        faceCentresVec3s,                                                               // cf
+        faceAreasVec3s,                                                                 // cn,
+        faceAreasVec3s,                                                                 // sf,
+        magFaceAreas,                                                                   // magSf,
+        faceAreasVec3s,                                                                 // nf,
+        {exec, {{-0.5, 0.0, 0.0}, {0.0, 0.5, 0.0}, {0.5, 0.0, 0.0}, {0.0, -0.5, 0.0}}}, // delta
+        {exec, {1, 1, 1, 1}},                                                           // weights
+        {exec, {2.0, 2.0, 2.0, 2.0}}, // deltaCoeffs --> mag(1 / delta)
+        {0, 1, 2, 3, 4}               // offset
+    );
+    return UnstructuredMesh(
+        {exec, {{0, 0, 0}, {0, 1, 0}, {1, 1, 0}, {1, 0, 0}}}, // points,
+        {exec, 1, 1.0},                                       // cellVolumes
+        {exec, {{0.5, 0.5, 0.0}}},                            // cellCentres
+        faceAreasVec3s,
+        faceCentresVec3s,
+        magFaceAreas,
+        {exec, {0, 0, 0, 0}}, // faceOwner
+        {exec, {}},           // faceNeighbour,
+        1,                    // nCells
+        0,                    // nInternalFaces,
+        4,                    // nBoundaryFaces,
+        4,                    // nBoundaries,
+        4,                    // nFaces,
+        boundaryMesh
+    );
+}
+
+} // namespace NeoN
