+++ test/mesh/unstructured/io/CMakeLists.txt
@@ -9,6 +9,8 @@ if(NeoN_WITH_VTK)
   neon_unit_test(cgnsBCPreservation)
   neon_unit_test(cgnsGeometricQuantities)
   neon_unit_test(cgnsMixedCells)
+  neon_unit_test(meshConnectivity)
+  neon_unit_test(meshGeometry)
   neon_unit_test(meshConverter)
   neon_unit_test(vtkHdfMeshWriter)
   neon_unit_test(vtkHdfMeshReader)
