+++ test/mesh/unstructured/io/tools/cgnsToVtu.cpp
@@ -8,7 +8,7 @@
 #include "NeoN/core/initialization.hpp"
 #include "NeoN/core/executor/serialExecutor.hpp"
 #include "NeoN/mesh/unstructured/io/cgnsMeshReader.hpp"
-#include "NeoN/mesh/unstructured/io/vtuMeshWriter.hpp"
+#include "NeoN/mesh/unstructured/io/vtmMeshWriter.hpp"
 
 int main(int argc, char* argv[])
 {
@@ -16,7 +16,7 @@ int main(int argc, char* argv[])
 
     if (argc < 3)
     {
-        std::cerr << "Usage: cgnsToVtu <input.cgns> <output.vtu>\n";
+        std::cerr << "Usage: cgnsToVtu <input.cgns> <output.vtm>\n";
         NeoN::finalize();
         return 1;
     }
@@ -24,7 +24,7 @@ int main(int argc, char* argv[])
     {
         NeoN::SerialExecutor exec;
         auto mesh = NeoN::io::readCgns(argv[1], exec);
-        NeoN::io::writeVtu(mesh, argv[2]);
+        NeoN::io::writeVtm(mesh, argv[2]);
 
         std::cout << "cells=" << mesh.nCells() << " points=" << mesh.points().size() << "\n";
     }
